tests/end2end/helpers: Move edit form calls to the Node class
